Update README.md Apologies for that first version 3.5. This one's much better
I'm not sure what I did wrong on the first version I think I compiled the wrong file. But it really wasn't too good anyway and this one I did some changes. It actually works about like I said it would because I'm testing it now as I'm publishing this note but I haven't uploaded it yet or compiled it. But it should be about version 3.8 or so.