Update README.md Version 1.7 works OK.
Version 1.7 actually works as it was designed. The previous version actually didn't do anything. It still moves the mouse across the screen yet it doesn't actually click on anything it just reads the area of the screen that's marked. Yet if you don't mark an area then it'll just simply read everything on the screen. Which could be useful for some reason but it takes a long time. Press escape to exit the program. The output is a text file in the same directory the program is run from.