does not move mouse  Default default scale 1 to 1
it still has the same real time display output the tracks each pixel with a cursor But it doesn't Take control of the mouse anymore. In the default scale is now 1 to 1 which is 100 times larger Which may be too large for most things, with these Low power introductory 3018's tiny milling machines move very slowly so even though it puts up a course facsimile,, it only takes it about an hour. I hope I will get it to go faster than that later