Update README.md
GCogen v1.9 .2 As the default set so that all you really have to do is scan anything and enter your scale factor if you need to or else it will default to one to one scale which drills a hole for every single pixel. I think it's set to one millimeter. But yet you can scale it to any scale you want. Of course it would be advisable that you would jog Z Axis so that it's much higher than the piece and then test it to see if it fits on the piece. Of course that's pending that whatever driver you're using, does not have a failure of its own. I used candle to develop this but you can use anything because it generates a standard G code block. Just like the other versions. The only difference is that this version has defaults that work. The previous versions had to be altered. And this version of course has its own cursor, so it doesn't actually move your mouse, but it does display what it's scanning. It's not listed in the instructions but if you push "Windows key"+"s", it displays the stats.        so you can See if you are getting a consistent scan over and over. Yet it does generate a file each time you scan and they can be quite large.